Split half of test_actor into test_actor_advanced (#7143)
